FoodHub/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
altgraph==0.17.4
    # via modulegraph
kybra==0.7.0
    # via -r requirements.in
modulegraph==0.19.3
    # via kybra
pycryptodome==3.21.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
